requirements.txt
asn1crypto==0.24.0
cffi==1.12.3
cryptography==2.7
cycler==0.10.0
Django==2.1
django-extensions==2.1.7
djangorestframework==3.9.4
kiwisolver==1.1.0
matplotlib==3.1.0
numpy==1.16.4
pyAesCrypt==0.4.3
pycparser==2.19
pycryptodome==3.8.2
pyparsing==2.4.0
python-crontab==2.3.6
python-dateutil==2.8.0
pytz==2019.1
six==1.12.0
sqlparse==0.3.0
